Add all new stats yaya

+++ GameJam/Assets/Scripts/Stats/Stat.cs
@@ -5,17 +5,17 @@ using UnityEngine;
 [System.Serializable]
 public class Stat
 {
-    [SerializeField]private int baseValue;
+    [SerializeField]private float baseValue;
 
-    private List<int> modifiers = new List<int>();
-    public int GetValue()
+    private List<float> modifiers = new List<float>();
+    public float GetValue()
     {
-        int finalValue = baseValue;
+        float finalValue = baseValue;
         modifiers.ForEach(x => finalValue += x);
         return finalValue;
     }
 
-    public void AddModifier(int modifier)
+    public void AddModifier(float modifier)
     {
         if (modifier != 0)
         {
@@ -23,7 +23,7 @@ public class Stat
         }
     }
 
-    public void RemoveModifier(int modifier)
+    public void RemoveModifier(float modifier)
     {
         if (modifier != 0)
         {
